+++ fastrace-datadog/src/lib.rs
@@ -36,27 +36,34 @@ impl DatadogReporter {
     fn convert<'a>(&'a self, spans: &'a [SpanRecord]) -> Vec<DatadogSpan<'a>> {
         spans
             .iter()
-            .map(move |s| DatadogSpan {
-                name: &s.name,
-                service: &self.service_name,
-                trace_type: &self.trace_type,
-                resource: &self.resource,
-                start: s.begin_time_unix_ns as i64,
-                duration: s.duration_ns as i64,
-                meta: if s.properties.is_empty() {
-                    None
-                } else {
-                    Some(
-                        s.properties
-                            .iter()
-                            .map(|(k, v)| (k.as_ref(), v.as_ref()))
-                            .collect(),
-                    )
-                },
-                error_code: 0,
-                span_id: s.span_id.0,
-                trace_id: s.trace_id.0 as u64,
-                parent_id: s.parent_id.0,
+            .map(move |s| {
+                let trace_id = s.trace_id.to_bytes();
+                DatadogSpan {
+                    name: &s.name,
+                    service: &self.service_name,
+                    trace_type: &self.trace_type,
+                    resource: &self.resource,
+                    start: s.begin_time_unix_ns as i64,
+                    duration: s.duration_ns as i64,
+                    meta: if s.properties.is_empty() {
+                        None
+                    } else {
+                        Some(
+                            s.properties
+                                .iter()
+                                .map(|(k, v)| (k.as_ref(), v.as_ref()))
+                                .collect(),
+                        )
+                    },
+                    error_code: 0,
+                    span_id: u64::from_be_bytes(s.span_id.to_bytes()),
+                    trace_id: u64::from_be_bytes(
+                        trace_id[8..].try_into().expect("trace id has 16 bytes"),
+                    ),
+                    parent_id: s
+                        .parent_id
+                        .map_or(0, |parent_id| u64::from_be_bytes(parent_id.to_bytes())),
+                }
             })
             .collect()
     }